Rename ZWave stubs so they are compatible with old iris code
