+++ agent/arcus-zw-controller/src/main/java/com/iris/agent/zw/ZWLocalProcessingDefault.java
@@ -24,7 +24,7 @@ import com.iris.protocol.zwave.Protocol;
 
 import rx.Observable;
 
-public class ZWLocalProcessingDefault implements ZWLocalProcessing {
+public class ZWaveLocalProcessingDefault implements ZWLocalProcessing {
    
    @Inject
    public ZWLocalProcessingDefault() {
