+++ agent/arcus-zw-controller/src/main/java/com/iris/agent/zw/ZWaveLocalProcessingNoop.java
@@ -18,14 +18,33 @@ package com.iris.agent.zw;
 import java.util.Collection;
 import java.util.concurrent.TimeUnit;
 
+import com.google.inject.Inject;
 import com.iris.messages.address.Address;
 import com.iris.protocol.zwave.Protocol;
 
 import rx.Observable;
 
-public interface ZWLocalProcessing {
-   boolean isOffline(Address addr);
-      void setOfflineTimeout(Address addr, long offlineTimeout);
-      Observable<?> send(Address addr, Protocol.Message msg);
-      void addScheduledPoll(Address addr, long period, TimeUnit unit, Collection<byte[]> payloads);
+public class ZWaveLocalProcessingNoop implements ZWLocalProcessing {
+   
+   @Inject
+   public ZWLocalProcessingNoop() {
+   }
+
+   @Override
+   public boolean isOffline(Address addr) {
+      return false;
+   }
+
+   @Override
+   public void setOfflineTimeout(Address addr, long offlineTimeout) {
+   }
+
+   @Override
+   public Observable<?> send(Address addr, Protocol.Message msg) {
+      return Observable.empty();
+   }
+
+   @Override
+   public void addScheduledPoll(Address addr, long period, TimeUnit unit, Collection<byte[]> payloads) {
+   }
 }
